ghagent/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
boto3==1.37.21
botocore==1.37.23
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
Deprecated==1.2.18
fastapi==0.115.12
googleapis-common-protos==1.70.0
grpcio==1.72.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib_metadata==8.6.1
InlineAgent @ file:///app/ghagent/AwsAgentApi/InlineAgent
jmespath==1.0.1
markdown-it-py==3.0.0
mcp==1.6.0
mdurl==0.1.2
openinference-semantic-conventions==0.1.16
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-exporter-otlp-proto-http==1.31.1
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
protobuf==5.29.5
pydantic==2.10.2
pydantic-settings==2.8.1
pydantic_core==2.27.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
requests==2.32.3
rich==13.9.4
s3transfer==0.11.5
six==1.17.0
sniffio==1.3.1
sse-starlette==2.3.6
starlette==0.46.2
termcolor==2.5.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
wrapt==1.17.2
zipp==3.22.0
